requirements.txt
beautifulsoup4==4.13.4
mlb_statsapi==1.9.0
openai==0.28.0
python-dotenv==1.1.1
Requests==2.32.4
selenium==4.34.2
